using generic version of the live mapper for telemetry data.
