+++ src/PitWallAcquisitionPlugin/PitWallAcquisitionPlugin.RunTime48.Tests/Aggregations/Aggregators/LiveAggregatorTest.cs
@@ -1,6 +1,6 @@
 ﻿using NFluent;
-using PitWallAcquisitionPlugin.Aggregations.Aggregators;
 using PitWallAcquisitionPlugin.Aggregations.Telemetries.Aggregators;
+using PitWallAcquisitionPlugin.Aggregations.Telemetries.Aggregators.Models;
 using PitWallAcquisitionPlugin.Tests.UI.ViewModels;
 using System;
 using System.Diagnostics;
