+++ src/PitWallAcquisitionPlugin/PitWallAcquisitionPlugin.RunTime48.Tests/WebApiForwarderServiceTest.cs
@@ -3,8 +3,8 @@ using FuelAssistantMobile.DataGathering.SimhubPlugin.Logging;
 using FuelAssistantMobile.DataGathering.SimhubPlugin.Repositories;
 using NFluent;
 using NSubstitute;
-using PitWallAcquisitionPlugin.Aggregations.Aggregators;
 using PitWallAcquisitionPlugin.Aggregations.Telemetries.Aggregators;
+using PitWallAcquisitionPlugin.Aggregations.Telemetries.Aggregators.Models;
 using System.Threading;
 using Xunit;
