+++ src/PitWallAcquisitionPlugin/PitWallAcquisitionPlugin/Aggregations/Telemetries/Mappers/LiveMapper.cs
@@ -7,31 +7,19 @@ namespace PitWallAcquisitionPlugin.Aggregations.Telemetries.Mappers
 
     public sealed class LiveMapper<TCounter> : ILiveMapper
     {
-        private readonly Func<IPluginRecordRepository, TCounter> sourceSelector;
-        private readonly Action<ITelemetryLiveAggregator, TCounter> setter;
+        private readonly GenericLiveMapper<TCounter, ITelemetryLiveAggregator> _mapper;
+        /**
+         * THOUGHT: We have some genericity here that could make this code reusable.
+         * */
 
         public LiveMapper(Func<IPluginRecordRepository, TCounter> sourceSelector, Action<ITelemetryLiveAggregator, TCounter> setter)
         {
-            this.sourceSelector = sourceSelector ?? throw new ArgumentNullException(nameof(sourceSelector));
-
-            this.setter = setter ?? throw new ArgumentNullException(nameof(setter));
+            _mapper = new GenericLiveMapper<TCounter, ITelemetryLiveAggregator>(sourceSelector, setter);
         }
 
         public void Set(IPluginRecordRepository adapter, ITelemetryLiveAggregator aggregator)
         {
-            if (adapter is null)
-            {
-                throw new ArgumentNullException(nameof(adapter));
-            }
-
-            if (aggregator is null)
-            {
-                throw new ArgumentNullException(nameof(aggregator));
-            }
-
-            var data = sourceSelector(adapter);
-
-            setter(aggregator, data);
+            _mapper.Set(adapter, aggregator);
         }
     }
 }
